requirements.txt
black==21.11b1
click==8.0.3
importlib-metadata==4.8.2
mcrcon==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.4.0
python-dotenv==0.19.2
regex==2021.11.10
tomli==1.2.2
typed-ast==1.5.0
typing-extensions==4.0.0
zipp==3.6.0
